+++ CHANGES.md
@@ -1,3 +1,8 @@
+@dev
+----
+* Added support for batch processing of multiple audio files
+* Added --verbose and muted most of the messages by default
+
 v1.0.0 (2018-10-01)
 -------------------
 * Initial release